CBD-5640: Only install snappy requirements.txt if present
Reviewed-on: https://review.couchbase.org/c/build-tools/+/198048
Tested-by: Blair Watt <blair.watt@couchbase.com>

+++ blackduck/couchbase-sdk-ruby/get_source.sh
@@ -35,7 +35,9 @@ function unset_bundle_path() {
     bundle config unset --local path
 }
 
-pip3 install -r ext/couchbase/third_party/snappy/third_party/benchmark/requirements.txt
+if [ -f "ext/couchbase/third_party/snappy/third_party/benchmark/requirements.txt" ]; then
+    pip3 install -r ext/couchbase/third_party/snappy/third_party/benchmark/requirements.txt
+fi
 
 bundle config set --local path './gems'
 bundle install
